+++ kodilla-basic-test/src/test/java/com/kodilla/basic_assertion/CalculatorTestSuit.java
@@ -23,23 +23,23 @@ public class CalculatorTestSuit {
     @Test
     public void testSquaredNegative () {
         Calculator calculator3 = new Calculator();
-        int a = -5;
-        int squaredResult = (int) calculator3.squared(a);
+        double a = -5.0;
+        double squaredResult = calculator3.squared(a);
         Assertions.assertEquals(25, squaredResult, 0.01);
 
     }
     @Test
     public void testSquaredZero() {
         Calculator calculator4 = new Calculator();
-        int a = 0;
-        int squaredResult = (int) calculator4.squared(a);
+        double a = 0.0;
+        double squaredResult = calculator4.squared(a);
         Assertions.assertEquals(0, squaredResult, 0.01);
     }
     @Test
     public void testSquaredPositive() {
         Calculator calculator5 = new Calculator ();
-        int a = 5;
-        int squaredResult = (int) calculator5.squared(a);
+        double a = 5.0;
+        double squaredResult = calculator5.squared(a);
         Assertions.assertEquals (25, squaredResult, 0.01);
     }
 }
